+++ core-ng/src/main/java/core/framework/api/util/ClasspathResources.java
@@ -1,7 +1,5 @@
 package core.framework.api.util;
 
-import core.framework.impl.io.ByteStreams;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UncheckedIOException;
@@ -34,7 +32,7 @@ public final class ClasspathResources {
             throw Exceptions.error("unexpected length of classpath resource, path={}, length={}", path, length);
 
         try (InputStream stream = connection.getInputStream()) {
-            return ByteStreams.readWithExpectedLength(stream, length);
+            return InputStreams.bytesWithExpectedLength(stream, length);
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }