+++ CHANGELOG.md
@@ -3,8 +3,14 @@
 * think about /test//b, and /:path(*)
 * refactory rabbitmq binding, unify
 * framework error, queue listener, background task error notification?
+* search require JSONMapper.toJSONValue()
 
 ## Change log
+### 4.1.6 (2/17/2016)
+* util: moved InputStreams to util, open to use
+* log: make max trace log 2000, and only append warn/error after
+* search: added SearchResponse<T> to convert hits to object on framework level
+
 ### 4.1.5 (2/16/2016)
 * tuning: internal tuning, move low level optimization class to impl package, leave core.framework.util simple ones
 * httpclient: use byte[] as body, remove ByteBuf